tests/wpt/tests/html/semantics/interactive-elements/the-dialog-element/dialog-inert.html

<!doctype html>
<meta charset=utf-8>
<link rel="author" title="Emilio Cobos Ãlvarez" href="mailto:emilio@crisal.io">
<link rel="author" title="Mozilla" href="https://mozilla.org">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<style>
  body { margin: 0 }
  dialog {
    width: 100%;
    height: 100%;
    max-width: 100%;
    max-height: 100%;
    box-sizing: border-box;
    padding: 0;
  }
  dialog::backdrop {
    display: none;
  }
</style>
<dialog id=dialog>Something</dialog>
<script>
test(function() {
  let dialog = document.getElementById("dialog");
  dialog.showModal();
  assert_equals(
    document.elementFromPoint(10, 10),
    dialog,
    "Dialog is hittable by default",
  );
  dialog.inert = true;
  assert_not_equals(
    document.elementFromPoint(10, 10),
    dialog,
    "Dialog becomes inert dynamically",
  );
  dialog.close();
  dialog.showModal();
  assert_not_equals(
    document.elementFromPoint(10, 10),
    dialog,
    "Dialog remains inert after open",
  );
});
</script>
